+++ nix/nixpkgs-version.nix
@@ -2,6 +2,6 @@
 # nix-prefetch-url --unpack https://github.com/NixOS/nixpkgs/archive/07723f14bb132f6d65c74b953bef39cd80457273.tar.gz
 
 {
-  rev    = "5bd2ca749d72461a4f2749f45e3ecd1084af5a44";
-  sha256 = "02jsids0mb1y196s9dzibhqcazfddywdx4i4z5yylrxb9yryjn0k";
+  rev    = "e4d4080db39749c9dde80c54b9ef486a8acf78a9";
+  sha256 = "1mcnnn3d7xxyh0zm6hvr9z8ai43n54hjw41vylhlx6jyw5kkxdxc";
 }